HUD上增加了罗盘 物理改进（每次都这么说但是可能越改越烂了）

+++ Aircraft.h
@@ -16,7 +16,7 @@ class Aircraft :
 	float controlx, controly;
 	float getLength(const glm::vec3&);
 	glm::vec3 getAcceleration();
-	int _getHDG() const;
+	float _getHDG() const;
 public:
 	Aircraft();
 	~Aircraft();
